// frontend/src/shared/constants/enums.ts
export enum CourseStatus {
  APPROVED = "approved",
  PENDING = "pending",
  REJECTED = "rejected",
}
export enum CourseLevel {
  BEGINNER = "beginner",
  INTERMEDIATE = "intermediate",
  ADVANCED = "advanced",
}
export enum CourseType {
  PAID = "paid",
  FREE = "free",
}

export enum QuizType {
  MULTIPLE_CHOICE = "multiple_choice",
  TRUE_FALSE = "true_false",
  SINGLE_CHOICE = "single_choice",
  FILL_IN_THE_BLANKS = "fill_in_the_blanks",
}

export enum CourseInfo {
  REQUIREMENTS = "requirements",
  QA = "qa",
  BENEFITS = "benefits",
  TECHNIQUES = "techniques",
  DOCUMENTS = "documents",
}
export enum OrderStatus {
  PENDING = "pending",
  COMPLETED = "completed",
  CANCELED = "canceled",
}
export enum CommentStatus {
  APPROVED = "approved",
  PENDING = "pending",
  REJECTED = "rejected",
}
export enum ReviewStatus {
  ACTIVE = "active",
  INACTIVE = "inactive",
}
export enum ReactionType {
  LIKE = "like",
  LOVE = "love",
  CARE = "care",
  FUN = "fun",
  WOW = "wow",
  SAD = "sad",
  ANGRY = "angry",
}

export enum CouponType {
  PERCENT = "percent",
  FIXED = "fixed",
}
export enum CouponStatus {
  ACTIVE = "active",
  INACTIVE = "inactive",
}
export enum UserStatus {
  ACTIVE = "active",
  INACTIVE = "inactive",
}
export enum UserRole {
  ADMIN = "admin",
  USER = "user",
  EXPERT = "expert",
}

export enum LessonType {
  VIDEO = "video",
  QUIZ = "quiz",
}